Fixed omissions in VncViewer for case of not inSeparateFrame
Small conditional changes made in:
fatalError(String, Exception)
fatalError(String)
disconnect()

These additions avoid System.exit(int) calls when VncViewer is being run
not as a separate frame.

This does not affect the default behavior of VncViewer as a stand-alone app
or when embedded as an Applet. The modifications are designed to make
VncViewer easier to embed in other Java projects.